requirements.txt
requests
python-dotenv
plotly
sendgrid==5.6.0